requirements.txt
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
Flask
gpib-ctypes
numpy
opencv-python
pdfkit
plotly
python-dotenv
PyVISA
PyVISA-py
PyYAML
scikit-image
scipy
SQLAlchemy
tensorflow